+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Save all translations during installation ([https://github.com/PierreRambaud/GotCms/issues/90](https://github.com/PierreRambaud/GotCms/issues/90))
 - Bug with Gc\Db\AbstractTable during installation ([https://github.com/PierreRambaud/GotCms/issues/91](https://github.com/PierreRambaud/GotCms/issues/91))
 - Add russian translation ([https://github.com/PierreRambaud/GotCms/issues/89](https://github.com/PierreRambaud/GotCms/issues/89))
+- Add Core\Config helper ([https://github.com/PierreRambaud/GotCms/issues/85](https://github.com/PierreRambaud/GotCms/issues/85))
 
 ##0.1.8 (02 Jun 2013):
 - Problem with replacement and mixed datatype ([https://github.com/PierreRambaud/GotCms/issues/55](https://github.com/PierreRambaud/GotCms/issues/55))